backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# CLIP and Vision Models
openai-clip==1.0.1
ftfy==6.1.3
regex==2023.10.3
tqdm==4.66.1

# Vector Database
qdrant-client==1.7.0

# Object Storage
minio==7.2.0
boto3==1.29.7

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-magic==0.4.27
ffmpeg-python==0.2.0

# API Documentation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0 